put args validation in a function + error handling in a new file + created enum for error types
